requirements.txt
Flask==1.1.2
pillow
requests
numpy
imgurpython